# apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CAMerlang API",
    "description": "CAMerlang API List",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://camerlang-api-service-hgtivvbcdq-uc.a.run.app",
      "description": ""
    }
  ],
...
  "paths": {
    "/articles": {
      "get": {
        "tags": ["Articles"],
        "summary": "Get All Articles",
        "description": "Get Articles Data",
        "responses": {
          "200": {
            "description": "success"
          },
          "404": {
            "description": "not found"
          }
        }
      },
      "post": {
        "tags": ["Articles"],
        "summary": "Post Article",
        "consumes": "multipart/form-data",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "thumbnail": {
                    "type": "string",
                    "format": "binary"
                  },
                  "title": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "read_duration": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/articles/{id}": {
      "get": {
        "tags": ["Articles"],
        "summary": "Get One Article",
        "description": "",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "description": "ID Article"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": ["Articles"],
        "summary": "Update One Article",
        "description": "",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "description": "ID Article"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "thumbnail": {
                    "type": "string",
                    "format": "binary"
                  },
                  "type": {
                    "type": "string"
                  },
                  "read_duration": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": ["Articles"],
        "summary": "Delete Article",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "description": "ID Article"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  }
}